// frontend/src/types/generated/GachaHistoryEnvelopeDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GachaPaginationInfoDto } from "./GachaPaginationInfoDto";
import type { GachaRecordEntryDto } from "./GachaRecordEntryDto";
import type { HistoryFiltersAppliedDto } from "./HistoryFiltersAppliedDto";

export type GachaHistoryEnvelopeDto = { records: Array<GachaRecordEntryDto>; pagination: GachaPaginationInfoDto; filtersApplied: HistoryFiltersAppliedDto };