// frontend/src/types/generated/GachaRecordsDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GachaTypeRecordsDto } from "./GachaTypeRecordsDto";

export type GachaRecordsDto = { limited: GachaTypeRecordsDto; regular: GachaTypeRecordsDto; special: GachaTypeRecordsDto };