// frontend/src/types/generated/HistoryFiltersAppliedDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DateRangeDto } from "./DateRangeDto";

export type HistoryFiltersAppliedDto = { rarity: number | null; gachaType: string | null; charId: string | null; dateRange: DateRangeDto | null };